Move test utils to their own file. (#132) (#135)
