+++ src/lib.rs
@@ -2,6 +2,9 @@ mod adapter;
 mod attributes;
 mod indexed_crate;
 
+#[cfg(test)]
+pub(crate) mod test_util;
+
 // Re-export the Crate type so we can deserialize it.
 pub use rustdoc_types::Crate;
